pyproject.toml
[project]
name = "d3-item-salvager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pydantic-settings>=2.10.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
    "typer>=0.16.0",
    "loguru>=0.7.3",
    "fastapi>=0.116.1",
    "dependency-injector>=4.48.1",
    "uvicorn>=0.35.0",
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "apscheduler>=3.11.0",
]

[dependency-groups]
dev = [
    "hatch>=1.14.1",
    "types-requests>=2.32.4.20250611",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "pytest-cov>=6.2.1",
    "prometheus-client>=0.22.1",
    "pre-commit>=4.2.0",
    "pyright>=1.1.403",
    "pylint>=3.3.7",
    "pre-commit-update>=0.8.0",
    "mypy>=1.17.0",
    "ruff>=0.12.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/d3_item_salvager"]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
addopts = ["-v", "--color=yes"]
testpaths = ["tests"]
python_files = "test_*.py"


# [tool.mypy]
# python_version = "3.12"
# mypy_path = "src"
# plugins = ['pydantic.mypy', 'sqlalchemy.ext.mypy.plugin']
# strict = true

# [tool.pylint.main]
# extension-pkg-allow-list = ["dependency_injector"]

[tool.pre-commit-update]
# dry_run = true
# dry_run_mode = "strict"
# all_versions = false
# verbose = true
# warnings = true
# preview = false
# jobs = 10
# exclude = ["isort"]
# keep = ["ruff"]
# bleeding_edge = ["pylint"]
# tag_prefix = [["black", "custom-tag-schema-v"]]

[tool.pyright]
typeCheckingMode = "strict"
